Feat miniprogram ci  (#172)
* feat: 修改miniprogram-ci

* feat: 修改miniprogram-ci上传appid

* feat: 修改node为ECMAScript模块

* feat: 测试

* feat: 测试

* feat: 测试

* feat: 修改workflows

---------

+++ build/ci.js
@@ -1,15 +1,21 @@
-const path = require('path')
-const ci = require('miniprogram-ci')
-const fs = require('fs')
-const packageJson = require('../package.json')
+import path from 'path'
+import { fileURLToPath } from 'url'
+import fs from 'fs'
+import ci from 'miniprogram-ci'
+import packageJson from '../package.json' with { type: 'json' }
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
+const privateKeyPath = path.resolve(__dirname, './key')
 
-const privateKeyContent = process.env.WX_PRIVATE_KEY
-if (!privateKeyContent) {
-  throw new Error('未找到私钥内容，请确保已正确配置 GitHub Secrets')
+// 检查私钥文件是否已存在
+if (!fs.existsSync(privateKeyPath)) {
+  const privateKeyContent = process.env.WX_PRIVATE_KEY
+  if (!privateKeyContent) {
+    throw new Error('未找到私钥内容，请确保已正确配置 GitHub Secrets')
+  }
+  console.log('>>>>写入私钥文件：', privateKeyPath);
+  fs.writeFileSync(privateKeyPath, privateKeyContent)
 }
-
-const privateKeyPath = path.resolve(__dirname, './private.key')
-fs.writeFileSync(privateKeyPath, privateKeyContent)
 
 const project = new ci.Project({
   appid: 'wxe5f52902cf4de896',
@@ -30,9 +36,11 @@ const params = {
     minifyWXML: true,
     minifyWXSS: true,
     codeProtect: false,
-    autoPrefixWXSS: true
+    autoPrefixWXSS: true,
+    ignoreUploadUnusedFiles: true
   },
 }
+await ci.packNpm(project, {})
 ci.upload({
   project,
   ...params
@@ -45,3 +53,5 @@ ci.upload({
   // 删除临时私钥文件
   fs.unlinkSync(privateKeyPath)
 })
+
+
